requirements.txt
python-dotenv~=1.0.1
langchain~=0.2.1
langchain-community~=0.2.1
huggingface-hub~=0.23.2
requests~=2.32.3
newspaper3k~=0.2.8
tiktoken~=0.7.0
lxml-html-clean~=0.1.1
pydantic~=2.7.2
streamlit~=1.35.0